move ArtifactComponentReference class to internal.io.reference package
