+++ components/uuf-core/src/main/java/org/wso2/carbon/uuf/internal/io/ArtifactFragmentReference.java
@@ -22,6 +22,7 @@ import org.wso2.carbon.uuf.api.reference.ComponentReference;
 import org.wso2.carbon.uuf.api.reference.FileReference;
 import org.wso2.carbon.uuf.api.reference.FragmentReference;
 import org.wso2.carbon.uuf.exception.UUFException;
+import org.wso2.carbon.uuf.internal.io.reference.ArtifactComponentReference;
 
 import java.nio.file.Files;
 import java.nio.file.Path;
