+++ components/uuf-core/src/main/java/org/wso2/carbon/uuf/internal/io/reference/ArtifactAppReference.java
@@ -23,7 +23,6 @@ import org.wso2.carbon.uuf.api.reference.ComponentReference;
 import org.wso2.carbon.uuf.api.reference.FileReference;
 import org.wso2.carbon.uuf.api.reference.ThemeReference;
 import org.wso2.carbon.uuf.exception.FileOperationException;
-import org.wso2.carbon.uuf.internal.io.ArtifactComponentReference;
 import org.wso2.carbon.uuf.internal.io.ArtifactFileReference;
 import org.wso2.carbon.uuf.internal.io.ArtifactThemeReference;
